run.sh runs
